+++ playground/nuxt.config.ts
@@ -1,6 +1,10 @@
 export default defineNuxtConfig({
   ssr: process.env.NUXT_SSR !== 'false',
 
+  future: {
+    compatibilityVersion: 4,
+  },
+
   devtools: {
     enabled: false,
   },
@@ -19,6 +23,19 @@ export default defineNuxtConfig({
     spaLoadingTemplate: {
       name: 'bar-scale',
     },
+    themeConfig: {
+      light: {
+        common: {
+          primaryColor: 'rgb(29, 78, 216)',
+        },
+      },
+      dark: {
+        common: {
+          bodyColor: 'rgb(16, 16, 20)',
+          primaryColor: 'rgb(96, 165, 250)',
+        },
+      },
+    },
   },
 
   tailwindcss: {
@@ -32,4 +49,6 @@ export default defineNuxtConfig({
       },
     },
   },
+
+  compatibilityDate: '2024-07-25',
 })