+++ src/runtime/components/NaiveConfig.vue
@@ -1,5 +1,6 @@
 <template>
   <n-config-provider
+    :key="key"
     :theme-overrides="naiveTheme"
     inline-theme-disabled
   >
@@ -10,40 +11,25 @@
 <script setup lang="ts">
 import type { ConfigProviderProps } from 'naive-ui'
 import { defu } from 'defu'
-import type { ThemeConfig, Theme, PublicConfig } from '../types'
+import type { Theme, PublicConfig } from '../types'
 import {
   useHead,
   useRuntimeConfig,
-  onMounted,
   useNaiveColorMode,
   useNaiveDevice,
+  computed,
+  ref,
+  onMounted,
   useNuxtApp,
-  useAppConfig,
-  watch,
-  useNuxtData,
 } from '#imports'
 
-interface NaiveConfigProps
-  extends /* @vue-ignore */ Omit<ConfigProviderProps, 'themeOverrides' | 'theme'> {
-  /** @deprecated since version 1.12.0, instead use `naiveui.themeConfig` in `app.config` */
-  themeConfig?: ThemeConfig
-}
+defineProps<Omit<ConfigProviderProps, 'themeOverrides' | 'theme'>>()
 
-const props = defineProps<NaiveConfigProps>()
 const { colorMode } = useNaiveColorMode()
-const config = useRuntimeConfig().public.naiveui as PublicConfig
-const appConfig = useAppConfig().naiveui as { themeConfig?: ThemeConfig } | undefined
-
-const themeConfig: ThemeConfig | undefined
-  = props.themeConfig
-  ?? appConfig?.themeConfig
-  ?? config.themeConfig
-
-const { data: naiveTheme } = useNuxtData<Theme>('naive-theme-config')
-
-naiveTheme.value ||= await updateTheme()
-
-watch(colorMode, () => updateTheme().then(t => (naiveTheme.value = t)))
+const runtimeConfig = useRuntimeConfig().public.naiveui as PublicConfig
+const deviceTheme = getDeviceTheme()
+const naiveTheme = computed(() => getTheme())
+const key = ref(1)
 
 useHead(() => ({
   htmlAttrs: {
@@ -66,75 +52,32 @@ onMounted(() => {
   const isPrerendered = typeof useNuxtApp().payload.prerenderedAt === 'number'
 
   if (isPrerendered && naiveTheme.value) {
-    // In order to update dom on pre-rendered pages
-    updateTheme().then(t => (naiveTheme.value = t))
+    key.value++
   }
 })
 
-async function updateTheme() {
-  const [deviceTheme, colorModeTheme] = await Promise.all([getDeviceTheme(), getColorModeTheme()])
-  const sharedTheme = evalTheme(themeConfig?.shared)
-  const theme = defu(sharedTheme, deviceTheme, colorModeTheme)
-  if (config.spaLoadingTemplate) {
+function getTheme() {
+  const colorModeTheme = colorMode.value === 'dark' ? runtimeConfig.themeConfig.dark : runtimeConfig.themeConfig.light
+  const theme = defu(runtimeConfig.themeConfig.shared, deviceTheme, colorModeTheme)
+  if (runtimeConfig.spaLoadingTemplate) {
     setLoadingTemplateTheme(theme)
   }
   return theme
 }
 
-async function getColorModeTheme() {
-  if (colorMode.value === 'dark') {
-    const darkTheme = evalTheme(themeConfig?.dark)
-    if (darkTheme?.defaults === false) {
-      return darkTheme
-    }
-    else {
-      const defaultDarkTheme = await import('../theme/dark')
-      return defu(darkTheme, defaultDarkTheme.default)
-    }
-  }
-  else {
-    const lightTheme = evalTheme(themeConfig?.light)
-    if (lightTheme?.defaults === false) {
-      return lightTheme
-    }
-    else {
-      const defaultLightTheme = await import('../theme/light')
-      return defu(lightTheme, defaultLightTheme.default)
-    }
-  }
-}
-
-async function getDeviceTheme() {
+function getDeviceTheme() {
   const { isMobileOrTablet, isMobile } = useNaiveDevice()
 
   if (isMobileOrTablet) {
-    const mobileOrTabletTheme = evalTheme(themeConfig?.mobileOrTablet)
-    if (mobileOrTabletTheme?.defaults === false) {
-      return mobileOrTabletTheme
-    }
-    else {
-      const defaultMobileOrTabletTheme = await import('../theme/mobileOrTablet')
-      return defu(mobileOrTabletTheme, defaultMobileOrTabletTheme.default)
-    }
+    return runtimeConfig.themeConfig.mobileOrTablet
   }
   else if (isMobile) {
-    const mobileTheme = evalTheme(themeConfig?.mobile)
-    if (mobileTheme?.defaults === false) {
-      return mobileTheme
-    }
-    else {
-      const defaultMobileOrTabletTheme = await import('../theme/mobileOrTablet')
-      return defu(mobileTheme, defaultMobileOrTabletTheme.default)
-    }
+    return defu(runtimeConfig.themeConfig.mobile, runtimeConfig.themeConfig.mobileOrTablet)
   }
 }
 
 function compileStyle(prop: string, value?: string) {
   return value && `${prop}: ${value} !important;`
-}
-
-function evalTheme<T extends object>(theme?: T | (() => T)) {
-  return typeof theme === 'function' ? theme() : theme
 }
 
 function setLoadingTemplateTheme(theme?: Theme) {